# openapi.json
{
  "components": {
    "schemas": {
      "ApproveUserRequest": {
        "description": "Approve a pending user, optionally also assigning them to\nchapters. The list may be empty \u2014 a freshly approved user\nwith no chapters lands on the dashboard's onboarding banner\nwhere they self-pick. The admin doesn't have to know which\nchapters apply at approval time, which matters when the\nperson who registered is the one with that context.",
        "properties": {
          "chapter_ids": {
            "items": {
              "type": "string"
            },
            "title": "Chapter Ids",
            "type": "array"
          }
        },
        "title": "ApproveUserRequest",
        "type": "object"
      },
...
      "ChapterRef": {
        "description": "Lightweight chapter reference embedded in ``UserOut``. Just\nenough for the frontend to render a chip without a second\nround-trip.",
        "properties": {
          "id": {
            "title": "Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name"
        ],
        "title": "ChapterRef",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "PendingCountOut": {
        "description": "Tiny DTO for the navbar's pending-approval indicator. Admin-\nonly \u2014 surfacing this to organisers would just be noise.",
        "properties": {
          "count": {
            "title": "Count",
            "type": "integer"
          }
        },
        "required": [
          "count"
        ],
        "title": "PendingCountOut",
        "type": "object"
      },
      "RenameUserRequest": {
        "description": "Min-length=1 catches empty strings; the handler additionally\nstrips whitespace and 422s on a name that's whitespace-only,\nsince Pydantic's ``min_length`` doesn't strip.",
        "properties": {
          "name": {
            "minLength": 1,
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "name"
        ],
        "title": "RenameUserRequest",
        "type": "object"
      },
      "SetUserChaptersRequest": {
        "description": "Replace the user's full chapter membership set. ``min_length=1``\nkeeps approved users in at least one chapter \u2014 emptying the set\nis the same as de-activating them, which is what the delete\nendpoint is for.",
        "properties": {
          "chapter_ids": {
            "items": {
              "type": "string"
            },
            "minItems": 1,
            "title": "Chapter Ids",
            "type": "array"
          }
        },
        "required": [
          "chapter_ids"
        ],
        "title": "SetUserChaptersRequest",
        "type": "object"
      },
...
      "UserOut": {
        "properties": {
          "chapters": {
            "items": {
              "$ref": "#/components/schemas/ChapterRef"
            },
            "title": "Chapters",
            "type": "array"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "email": {
            "title": "Email",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "is_approved": {
            "title": "Is Approved",
            "type": "boolean"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "role": {
            "title": "Role",
            "type": "string"
          }
        },
        "required": [
          "id",
          "email",
          "name",
          "role",
          "is_approved",
          "chapters",
          "created_at"
        ],
        "title": "UserOut",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "Opkomst",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/admin/users": {
      "get": {
        "description": "Open to every approved user. Returns the same projection\neveryone else sees \u2014 names + email + chapter memberships.\nNo \"see only your own row\" mode; the visibility matches the\norganising context the project is built for.",
        "operationId": "list_users_api_v1_admin_users_get",
        "parameters": [
          {
            "in": "query",
            "name": "pending",
            "required": false,
            "schema": {
              "default": false,
              "title": "Pending",
              "type": "boolean"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/UserOut"
                  },
                  "title": "Response List Users Api V1 Admin Users Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Users",
        "tags": [
          "admin"
        ]
      }
    },
    "/api/v1/admin/users/pending-count": {
      "get": {
        "description": "How many live users are awaiting admin approval. Backs the\nred-dot indicator on the navbar's Accounts link \u2014 only\nrendered for admins, and the endpoint is admin-only too so\norganisers can't sniff the count.\n\nDefined *before* ``GET /users`` so the static path beats\nFastAPI's parametric ``/users/{user_id}/...`` patterns; the\nexplicit string segment wouldn't actually conflict with the\nlist endpoint, but keeping all the static-segment routes up\ntop is the cheapest way to never trip over that ordering rule\nin the future.",
        "operationId": "pending_user_count_api_v1_admin_users_pending_count_get",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PendingCountOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Pending User Count",
        "tags": [
          "admin"
        ]
      }
    },
    "/api/v1/admin/users/{user_id}": {
      "delete": {
        "description": "Soft-delete a user. Stamps ``deleted_at``; the email slot\nfrees up via the partial-unique index, so registering again\nwith the same address restores the row in an unapproved\nstate.\n\nSelf-deletion is implicitly blocked because the permission\nmatrix's DELETE_USER branch is admin-only and contains no\nself-service carve-out \u2014 an admin trying to delete their\nown row therefore goes through the matrix and is rejected\nby the explicit \"not self\" check we still keep here as\ndefence in depth (and to give a clearer 409 over the\nmatrix's blanket 403).",
        "operationId": "delete_user_api_v1_admin_users__user_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "user_id",
            "required": true,
            "schema": {
              "title": "User Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete User",
        "tags": [
          "admin"
        ]
      }
    },
    "/api/v1/admin/users/{user_id}/approve": {
      "post": {
        "description": "Mark a pending user approved and replace their membership\nset with ``chapter_ids``. The approval email lists every\nassigned chapter so the user knows the full scope of what\nthey've been let into.",
        "operationId": "approve_user_api_v1_admin_users__user_id__approve_post",
        "parameters": [
          {
            "in": "path",
            "name": "user_id",
            "required": true,
            "schema": {
              "title": "User Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApproveUserRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Approve User",
        "tags": [
          "admin"
        ]
      }
    },
    "/api/v1/admin/users/{user_id}/demote": {
      "post": {
        "description": "Demote an admin back to organiser. The matrix blocks\nself-demotion \u2014 the same anti-foot-shoot rule that previously\nlived inline in this handler.",
        "operationId": "demote_user_api_v1_admin_users__user_id__demote_post",
        "parameters": [
          {
            "in": "path",
            "name": "user_id",
            "required": true,
            "schema": {
              "title": "User Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Demote User",
        "tags": [
          "admin"
        ]
      }
    },
    "/api/v1/admin/users/{user_id}/promote": {
      "post": {
        "operationId": "promote_user_api_v1_admin_users__user_id__promote_post",
        "parameters": [
          {
            "in": "path",
            "name": "user_id",
            "required": true,
            "schema": {
              "title": "User Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Promote User",
        "tags": [
          "admin"
        ]
      }
    },
    "/api/v1/admin/users/{user_id}/rename": {
      "post": {
        "description": "Update a user's display name. Self-service. The user-\nsupplied value is stripped of surrounding whitespace; an\nempty-after-strip name is rejected before the row is touched\n(Pydantic's ``min_length`` doesn't strip).",
        "operationId": "rename_user_api_v1_admin_users__user_id__rename_post",
        "parameters": [
          {
            "in": "path",
            "name": "user_id",
            "required": true,
            "schema": {
              "title": "User Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RenameUserRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Rename User",
        "tags": [
          "admin"
        ]
      }
    },
    "/api/v1/admin/users/{user_id}/set-chapters": {
      "post": {
        "description": "Replace the user's full chapter membership set. Self-service:\na user can manage their own chapter memberships without\nadmin intervention. ``min_length=1`` on the schema keeps\nevery approved user in at least one chapter \u2014 emptying the\nset would orphan you from the project.",
        "operationId": "set_user_chapters_api_v1_admin_users__user_id__set_chapters_post",
        "parameters": [
          {
            "in": "path",
            "name": "user_id",
            "required": true,
            "schema": {
              "title": "User Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetUserChaptersRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Set User Chapters",
        "tags": [
          "admin"
        ]
      }
    },
...
  }
}